Move proxyHandler* to their own file
